+++ Mesh/highOrderSmoother.cpp
@@ -328,7 +328,7 @@ void addOneLayer(const std::vector<MElement*> &v,
 void _printJacobiansAtNodes (const char * name, std::vector<MElement*> &v)
 {
   FILE *fd = fopen (name,"w");
-  fprintf(fd,"$MeshFormat\n2 0 8\n$EndMeshFormat\n$ElementNodeData\n1\n\"det J\"\n1\n0.0\n3\n1\n1\n%d\n",v.size());
+  fprintf(fd,"$MeshFormat\n2 0 8\n$EndMeshFormat\n$ElementNodeData\n1\n\"det J\"\n1\n0.0\n3\n1\n1\n%d\n", (int) v.size());
   for (int i=0; i<v.size(); i++){
     const polynomialBasis* pb = v[i]->getFunctionSpace();
     fprintf(fd,"%d %d",v[i]->getNum(),v[i]->getNumVertices());
@@ -723,7 +723,7 @@ double highOrderSmoother::apply_incremental_displacement (double max_incr,
   
   //+++++++++ Move vertices @ maximum ++++++++++++++++++++++++++++++++++++
   FILE *fd = fopen ("d.msh","w");
-  fprintf(fd,"$MeshFormat\n2 0 8\n$EndMeshFormat\n$NodeData\n1\n\"tr(sigma)\"\n1\n0.0\n3\n1\n3\n%d\n",_vertices.size());
+  fprintf(fd,"$MeshFormat\n2 0 8\n$EndMeshFormat\n$NodeData\n1\n\"tr(sigma)\"\n1\n0.0\n3\n1\n3\n%d\n", (int) _vertices.size());
   for (std::set<MVertex*>::iterator it = _vertices.begin(); it != _vertices.end(); ++it){
     double ax, ay, az;
     myAssembler.getDofValue(*it, 0, getTag(), ax);
